Simplify tooltip filter compilation and restrict filter to tooltip arrays
- Reuse existing isFieldPredicate/isFieldValidPredicate guards instead of
  bespoke runtime validation that duplicated the schema
- Drop the equal-null special case (fieldFilterExpression handles null)
- Move TooltipFieldPredicate/TooltipFieldFilter next to the other predicate
  types in predicate.ts
- Remove TooltipFieldDefWithCondition: filter on a single (non-array) tooltip
  field def was silently ignored, so only allow it in the array form
- Replace the O(n^2) aria separator tracking with a leading separator plus
  slice()
- Emit a single ternary instead of merge() for single-field tooltips
- Remove the now-unused tooltipData
