+++ examples/compiled/bar_tooltip_filter.vg.json
@@ -81,7 +81,7 @@
           "fill": {"scale": "color", "field": "nonZeroTooltip"},
           "ariaRoleDescription": {"value": "bar"},
           "description": {
-            "signal": "\"category: \" + (isValid(datum[\"category\"]) ? isArray(datum[\"category\"]) ? join(datum[\"category\"], ' ') : datum[\"category\"] : \"\"+datum[\"category\"]) + \"; \" + \"plotValue: \" + (format(datum[\"plotValue\"], \"\")) + \"; \" + \"nonZeroTooltip: \" + (format(datum[\"nonZeroTooltip\"], \"\")) + ((isValid(datum[\"validTooltip\"]) && isFinite(+datum[\"validTooltip\"])) ? (\"; \" + \"validTooltip: \" + (format(datum[\"validTooltip\"], \"\"))) : \"\")"
+            "signal": "slice(\"; category: \" + (isValid(datum[\"category\"]) ? isArray(datum[\"category\"]) ? join(datum[\"category\"], ' ') : datum[\"category\"] : \"\"+datum[\"category\"]) + \"; plotValue: \" + (format(datum[\"plotValue\"], \"\")) + \"; nonZeroTooltip: \" + (format(datum[\"nonZeroTooltip\"], \"\")) + ((isValid(datum[\"validTooltip\"]) && isFinite(+datum[\"validTooltip\"])) ? \"; validTooltip: \" + (format(datum[\"validTooltip\"], \"\")) : \"\"), 2)"
           },
           "x": {"scale": "x", "field": "category"},
           "width": {"signal": "max(0.25, bandwidth('x'))"},
