+++ src/encoding.ts
@@ -85,7 +85,6 @@ import {
   TextDef,
   TimeDef,
   TooltipFieldDef,
-  TooltipFieldDefWithCondition,
   title,
   TypedFieldDef,
   vgField,
@@ -304,7 +303,7 @@ export interface Encoding<F extends Field> {
    *
    * See the [`tooltip`](https://vega.github.io/vega-lite/docs/tooltip.html) documentation for a detailed discussion about tooltip in Vega-Lite.
    */
-  tooltip?: TooltipFieldDefWithCondition<F> | StringValueDefWithCondition<F> | TooltipFieldDef<F>[] | null;
+  tooltip?: StringFieldDefWithCondition<F> | StringValueDefWithCondition<F> | TooltipFieldDef<F>[] | null;
 
   /**
    * A URL to load upon mouse click.
